rename cardVariants file to camelCase
