+++ src/components/Card/cardVariants.ts
